% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{Acos}
\alias{Acos}
\title{Compute the arccosine of x}
\usage{
Acos(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector containing the arccosine of each element of x.
}
\description{
This function computes the arccosine of the input value x.
}
\examples{
Acos(0)
}
