% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{Asin}
\alias{Asin}
\title{Compute the arcsine of x}
\usage{
Asin(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector containing the arcsine of each element of x.
}
\description{
This function computes the arcsine of the input value x.
}
\examples{
Asin(0)
}
