% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{Atan}
\alias{Atan}
\title{Compute the arc tangent of y/x}
\usage{
Atan(y, x)
}
\arguments{
\item{y}{Numeric vector, the y-coordinate.}

\item{x}{Numeric vector, the x-coordinate.}
}
\value{
Numeric vector containing the arctangent of y/x.
}
\description{
This function computes the arc tangent of the input values y and x.
}
\examples{
Atan(0, 1)
}
