% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{calculate_loss}
\alias{calculate_loss}
\title{Calculate Loss Function}
\usage{
calculate_loss(y, gamma)
}
\arguments{
\item{y}{Matrix of observed values.}

\item{gamma}{Matrix of predicted values.}
}
\value{
Loss value.
}
\description{
This function calculates the loss function based on the squared spherical distances between observed values and predicted values on the curve.
}
