% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{cross}
\alias{cross}
\title{Compute the cross product of two vectors}
\usage{
cross(u, v, normalize = FALSE)
}
\arguments{
\item{u}{Numeric vector.}

\item{v}{Numeric vector.}

\item{normalize}{logical. If TRUE, returns the normalized vector of the cross product result.}
}
\value{
Numeric vector representing the cross product of u and v.
}
\description{
This function computes the cross product of two input vectors u and v.
}
\examples{
cross(c(1,0,0), c(0,1,0))
}
