% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{cross_normalized}
\alias{cross_normalized}
\title{Compute the normalized cross product of two vectors}
\usage{
cross_normalized(u, v)
}
\arguments{
\item{u}{Numeric vector.}

\item{v}{Numeric vector.}
}
\value{
Numeric vector representing the normalized cross product of u and v.
}
\description{
This function computes the cross product of two input vectors u and v,
normalizes the result, and returns the normalized vector.
}
\examples{
cross_normalized(c(1,0,0), c(0,1,0))
}
