% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{dot}
\alias{dot}
\title{Compute the dot product of two vectors}
\usage{
dot(u, v)
}
\arguments{
\item{u}{Numeric vector.}

\item{v}{Numeric vector.}
}
\value{
Numeric value representing the dot product of u and v.
}
\description{
This function computes the dot product of two input vectors u and v.
}
\examples{
dot(c(1,2,3), c(4,5,6))
}
