% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{edp}
\alias{edp}
\title{Compute the equal-distance projection of a point onto the xy plane}
\usage{
edp(p)
}
\arguments{
\item{p}{Numeric vector representing a point in Cartesian coordinates.}
}
\value{
Numeric vector representing the equal-distance projection of p onto the xy plane.
}
\description{
This function computes the equal-distance projection of a point p onto the xy plane.
}
\examples{
edp(c(1,1,1))
}
