% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{geodesic_lower}
\alias{geodesic_lower}
\title{Compute the value of the geodesic curve connecting two points on the unit sphere for a given time point t}
\usage{
geodesic_lower(t, p, q, a, b)
}
\arguments{
\item{t}{Time parameter for the geodesic path.}

\item{p}{Numeric vector representing the starting point.}

\item{q}{Numeric vector representing the ending point.}

\item{a}{Start time parameter.}

\item{b}{End time parameter.}
}
\value{
Numeric vector representing a point along the geodesic path.
}
\description{
This function computes points along the geodesic connecting two points p and q on the unit sphere.
}
\examples{
geodesic_lower(0.5, c(1,0,0), c(0,1,0), 0, 1)
}
