% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goni_spherical}
\alias{goni_spherical}
\title{A tropical cyclone dataset}
\description{
A tropical cyclone dataset provided by the Regional Specialized Meteorological Center (RSMC) Tokyo Typhoon Center. We select a cyclone called 'Goni' observed over time  on August, 2015. The first column represents the time points, and the remaining two columns provides the observed spherical coordinates.
}
