% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{jump_linear}
\alias{jump_linear}
\title{Calculate Jump Vector for Linear Spline}
\usage{
jump_linear(control_points, knots)
}
\arguments{
\item{control_points}{Matrix of control points.}

\item{knots}{Vector of knots.}
}
\value{
Jump vector.
}
\description{
This function calculates the jump vector for a linear spline model.
}
