% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{knots_quantile}
\alias{knots_quantile}
\title{Generate knots for the piecewise gedeosic curve based on the quantiles.}
\usage{
knots_quantile(x, dimension, tiny = 1e-05)
}
\arguments{
\item{x}{Numeric vector representing time points for the geodesic path.}

\item{dimension}{Numeric vector representing the starting point.}

\item{tiny}{Numeric vector representing the ending point.}
}
\value{
Numeric vector representing knots sequence in the time domain.
}
\description{
This function computes points along the geodesic connecting two points p and q on a Riemannian manifold at specified time points.
}
\examples{
knots_quantile(seq(0, 1, length.out = 100), 10)
}
