% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{norm2}
\alias{norm2}
\title{Compute the L2 norm (Euclidean norm) of a vector}
\usage{
norm2(u)
}
\arguments{
\item{u}{Numeric vector.}
}
\value{
Numeric value representing the L2 norm of u.
}
\description{
This function computes the L2 norm (Euclidean norm) of the input vector u.
}
\examples{
norm2(c(1,2,3))
}
