% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{normalize}
\alias{normalize}
\title{Normalize a matrix row-wise}
\usage{
normalize(x)
}
\arguments{
\item{x}{Numeric matrix.}
}
\value{
Numeric matrix with normalized rows.
}
\description{
This function normalizes the rows of the input matrix x by dividing each row by its L2 norm (Euclidean norm).
}
\examples{
normalize(matrix(c(1,2,3,4,5,6), nrow = 2, byrow = TRUE))
}
