% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{normalize_lower}
\alias{normalize_lower}
\title{This function normalizes the the input vector v by dividing its L2 norm (Euclidean norm).}
\usage{
normalize_lower(v)
}
\arguments{
\item{v}{Numeric vector.}
}
\value{
Numeric vector with normalized.
}
\description{
This function normalizes the the input vector v by dividing its L2 norm (Euclidean norm).
}
\examples{
normalize_lower(1:6)
}
