% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{omega}
\alias{omega}
\title{Compute the omega vector given spherical coordinates}
\usage{
omega(theta, phi)
}
\arguments{
\item{theta}{Numeric value representing the polar angle.}

\item{phi}{Numeric value representing the azimuthal angle.}
}
\value{
Numeric vector representing the omega vector.
}
\description{
This function computes the omega vector given spherical coordinates theta and phi.
}
\examples{
omega(pi/4, pi/3)
}
