% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{R_gradient_penalty}
\alias{R_gradient_penalty}
\title{Compute the Riemannian gradient penalty}
\usage{
R_gradient_penalty(control_points, knots, index)
}
\arguments{
\item{control_points}{Matrix representing the control points.}

\item{knots}{Numeric vector representing the knots.}

\item{index}{Index for computing the Riemannian gradient penalty.}
}
\value{
A list containing Riemannian gradient and the exponential map of the negative Riemannian gradient.
}
\description{
This function computes the Riemannian gradient penalty given control points, knots, and index.
}
\examples{
R_gradient_penalty(matrix(c(1,0,0,0,1,0,0,0,1), 3), c(0, 1, 2), 2)
}
