% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{restrict}
\alias{restrict}
\title{Restrict a value to a specified range}
\usage{
restrict(x, lower, upper)
}
\arguments{
\item{x}{Numeric value to be restricted.}

\item{lower}{Numeric value representing the lower bound of the range.}

\item{upper}{Numeric value representing the upper bound of the range.}
}
\value{
Numeric value restricted to the specified range.
}
\description{
This function restricts the input value x to the specified range [lower, upper].
}
\examples{
restrict(5, 0, 10)
}
