% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{Rgradient_loss_linear_spline}
\alias{Rgradient_loss_linear_spline}
\title{Calculate the Gradient of Loss Function for Linear Spline Curve}
\usage{
Rgradient_loss_linear_spline(y, t, control_points, knots, index)
}
\arguments{
\item{y}{Matrix of observed values.}

\item{t}{Vector of time or location values.}

\item{control_points}{Matrix of control points.}

\item{knots}{Vector of knots.}

\item{index}{Index of the control point.}
}
\value{
Matrix containing the gradient of the loss function.
}
\description{
This function calculates the gradient of the loss function for a linear spline curve.
}
