% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{spherical_dist}
\alias{spherical_dist}
\title{Calculate spherical distance between two vectors}
\usage{
spherical_dist(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
The distance between vectors x and y.
}
\description{
This function calculates the spherical distance between two vectors.
}
\examples{
x <- c(1, 0, 0)
y <- c(0, 1, 0)
spherical_dist(x, y)
}
