% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{Exp}
\alias{Exp}
\title{Compute the exponential map on a Riemannian manifold}
\usage{
Exp(x, v)
}
\arguments{
\item{x}{Numeric vector representing the base point.}

\item{v}{Numeric vector representing the tangent vector.}
}
\value{
Numeric vector representing the result of the exponential map.
}
\description{
This function computes the exponential map on a Riemannian manifold given a base point x and a tangent vector v.
}
\examples{
Exp(c(1,0,0), c(0,1,0))
}
