% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{Normalize}
\alias{Normalize}
\title{Normalize a matrix row-wise}
\usage{
Normalize(x)
}
\arguments{
\item{x}{Numeric matrix.}
}
\value{
Numeric matrix with normalized rows.
}
\description{
This function normalizes the rows of the input matrix x by applying the normalize function to each row.
}
\examples{
Normalize(matrix(c(1,2,3,4,5,6), nrow = 2, byrow = TRUE))
}
