% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-spherical-spline.R
\name{penalized_linear_spherical_spline}
\alias{penalized_linear_spherical_spline}
\title{Penalized Linear Spherical Spline}
\usage{
penalized_linear_spherical_spline(
  t,
  y,
  initial_control_points = NULL,
  dimension,
  initial_knots,
  lambdas,
  step_size = 0.01,
  maxiter = 1000,
  epsilon_iter = 1e-05,
  jump_eps = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{t}{A numeric vector representing the time or location.}

\item{y}{A matrix where each row represents a data point.}

\item{initial_control_points}{An optional matrix specifying initial control points. Default is NULL.}

\item{dimension}{An integer specifying the dimension of the spline.}

\item{initial_knots}{An optional numeric vector specifying initial knots. Default is NULL.}

\item{lambdas}{A numeric vector specifying the penalization parameters.}

\item{step_size}{A numeric value specifying the step size for optimization. Default is 0.01.}

\item{maxiter}{An integer specifying the maximum number of iterations. Default is 1000.}

\item{epsilon_iter}{A numeric value specifying the convergence criterion for iterations. Default is 1e-05.}

\item{jump_eps}{A numeric value specifying the threshold for pruning control points based on jump size. Default is 1e-02.}

\item{verbose}{A logical value indicating whether to print progress information. Default is FALSE.}
}
\value{
A list containing the fit information.
}
\description{
This function fits a penalized linear spherical spline to the given data.
}
\details{
This function estimates the optimal control points and knots for the given data, fits the model, and returns the optimized result.
Internally, gradient descent is used to minimize the loss of a given model, and a penalty term is added to control the complexity of the model.
Additionally, the BIC (Bayesian Information Criterion) value is calculated according to the model's complexity to provide information for model selection.
The function constructs piecewise curves using the piecewise_geodesic function and employs penalty terms to control the complexity of the model by updating control points and knots.
Its purpose is to find the optimal linear piecewise spline for the given data while controlling model complexity through penalty terms.
}
