\docType{package}
\name{spheresmooth-package}
\alias{spheresmooth}
\alias{spheresmooth-package}
\title{spheresmooth: Piecewise Geodesic Smoothing for Spherical Data}
\description{
This package offers a method of fitting a smooth path to a given set of noisy spherical data observed at known time. It implements a piecewise geodesic curve fitting method on the unit sphere based on a velocity-based penalization scheme. The proposed approach is implemented using the Riemannian block coordinate descent algorithm. To understand the method and algorithm, you can refer to Bak, K. Y., Shin, J. K., & Koo, J. Y. (2023) <\href{https://www.tandfonline.com/doi/full/10.1080/02664763.2022.2054962}{doi:10.1080/02664763.2022.2054962}> for the case of order 1. Additionally, this package includes various functions necessary for handling spherical data.
}
\author{
\strong{Maintainer}: Kwan-Young Bak \email{kybak@sungshin.ac.kr} (\href{https://orcid.org/0000-0002-4541-160X}{ORCID:0000-0002-4541-160X}) (Sungshin Women's University)

Authors:
\itemize{
  \item Seyoung Lee \email{20210861@sungshin.ac.kr} (Sungshin Women's University)
}

}
\keyword{internal}
